Adds Pretty Printed Source Code To Docs
Plugin generates pages using google prettify to highlight the source.
The three css files are the defaults that come with google prettify. The
plugin is set up to use the desert style as it's default stylesheet,
with prettify, sunburst and prettify-jsdoc listed as alternative
stylesheets.

Currently this plugin is not tied in with any template. There won't be
links in the documentation to bring you to the prettified source.